pyproject.toml
[project]
name = "simple-agent-with-skills"
version = "0.1.0"
description = "Minimal async Anthropic agent with tool use and skills"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
]

[project.scripts]
simple-agent-with-skills = "simple_agent_with_skills.cli:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
